+++ Library/Color.cs
@@ -1,0 +1,12 @@
+namespace OsuLoader {
+    public struct Color
+    {
+        public int R, G, B;
+        public Color(int r, int g, int b)
+        {
+            R = r;
+            G = g;
+            B = b;
+        }
+    }
+}